Move JNI loading to the classes with native methods.
