+++ trunk/video-overlay/src/main/java/org/mitre/mpf/frameextractor/FrameExtractor.java
@@ -29,6 +29,7 @@ package org.mitre.mpf.frameextractor;
 import org.javasimon.SimonManager;
 import org.javasimon.Split;
 import org.mitre.mpf.interop.JsonDetectionOutputObject;
+import org.mitre.mpf.videooverlay.JniLoader;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import com.google.common.collect.HashBasedTable;
@@ -47,6 +48,10 @@ import java.util.TreeMap;
 
 public class FrameExtractor {
     private static final Logger log = LoggerFactory.getLogger(FrameExtractor.class);
+
+    static {
+        JniLoader.ensureLoaded();
+    }
 
     private final URI media;
     private final Map<String, String> mediaMetadata;
