+++ trunk/video-overlay/src/main/java/org/mitre/mpf/videooverlay/BoundingBoxWriter.java
@@ -28,6 +28,10 @@ package org.mitre.mpf.videooverlay;
 
 public class BoundingBoxWriter {
 
+    static {
+        JniLoader.ensureLoaded();
+    }
+
     private BoundingBoxWriter() {
     }
